Put verilator outputs in a separate directory.
Verilator can output additional .cpp files if the verilog sources
contain code generation. Those files don't have a fixed name, so we
can't refer to them directly in the makefile. The solution is to put
all files generated by Verilator in a directory and compile all .cpp
files in that directory.